Address review: fold access benches into variant_validation, trim comments
Move the field-name access benchmarks into the existing `variant_validation`
bench instead of adding a separate binary, and cut the commentary around the
`debug_assert` and the SAFETY block in `VariantMetadata::get` down to what the
reader needs.
